src/auth-service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "Auth service to generate and validate tokens"
authors = ["aarav <aarav26r@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.3"}
pyjwt = "^2.8.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
sqlalchemy = {extras = ["mypy"], version = "^2.0.31"}
psycopg2-binary = "^2.9.9"
sqlalchemy-stubs = "^0.4"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"
black = "^24.4.2"
mypy = "^1.11.1"

[tool.black]
line-length = 88

[tool.ruff]
select = ["E", "F", "I"]
fixable = ["ALL"]
exclude = [".git", ".mypy_cache", ".ruff_cache"]
line-length = 88

[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
disallow_untyped_decorators = false
ignore_missing_imports = true
implicit_reexport = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
